# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402-browser-bridge",
    "version": "0.1.0",
    "description": "Turns any website flow into an x402-paid API — declare a flow in YAML, Puppeteer executes it, and the extracted data, screenshots and step log come back in the response body. Pay in USDC on Base or Solana; the client picks the rail. WARNING: automating a website you do not own may breach its terms of service. The bundled flows target pages this repository serves itself; the operator of any deployment is responsible for the legality of every flow it serves.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "contact": {
      "name": "nirholas",
      "email": "nichxbt@gmail.com",
      "url": "https://github.com/nirholas/x402-browser-bridge"
    },
    "x-guidance": "Framework that turns any website flow into an x402-paid API — declare a flow in YAML, Puppeteer executes, artifact returned in-response. Every paid route answers an unpaid request with HTTP 402 and a list of accepted payments (USDC on Base or Solana — pick either), then returns the purchased artifact in the body of the same request once you retry with an X-PAYMENT header. Payment and delivery happen in one round trip; nothing is promised for later. POST /run/:flow ($0.05) returns {extracted data, confirmation screenshot base64, step log} GET /flows (free) returns declared flows + their prices"
  },
  "servers": [
    {
      "url": "http://localhost:4034",
      "description": "Local dev server"
    }
  ],
  "tags": [
    {
      "name": "paid",
      "description": "Routes behind the x402 dual-rail paywall"
    },
    {
      "name": "free",
      "description": "Free routes, including the pre-payment validation failures"
    }
  ],
  "paths": {
    "/run/{flow}": {
      "post": {
        "tags": [
          "paid"
        ],
        "operationId": "runFlow",
        "summary": "Execute a declared browser flow and return its artifact",
        "description": "Price comes from the flow definition (bundled flows: $0.05 USDC). Returns 200 with a signed run record whether the flow completed or failed — on failure the step log and failure screenshot are the artifact. Unknown flow (404), invalid inputs (400) and no browser (503) are checked on the X-PAYMENT retry, before the payment is verified or settled, so those cost nothing. An unpaid request always receives the 402 challenge first — including for a flow name that does not exist — so the route stays discoverable.",
        "parameters": [
          {
            "name": "flow",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "examples": {
              "booking": {
                "value": "demo-booking"
              },
              "catalog": {
                "value": "demo-catalog"
              }
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "The flow's declared inputs. GET /flows for the schema.",
                "additionalProperties": true
              },
              "examples": {
                "booking": {
                  "value": {
                    "name": "Ada Lovelace",
                    "email": "ada@example.com",
                    "date": "2026-09-01",
                    "partySize": "4"
                  }
                },
                "booking-failure-path": {
                  "value": {
                    "name": "unavailable",
                    "email": "x@example.com",
                    "date": "2026-09-01"
                  }
                },
                "catalog": {
                  "value": {
                    "query": "bolt"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed run record. `status` is `completed` or `failed`; both are paid-for artifacts.",
            "headers": {
              "X-PAYMENT-RESPONSE": {
                "$ref": "#/components/headers/XPaymentResponse"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "run"
                  ],
                  "properties": {
                    "run": {
                      "$ref": "#/components/schemas/SignedRun"
                    },
                    "flow": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "price": {
                          "type": "string"
                        }
                      }
                    },
                    "paidWith": {
                      "$ref": "#/components/schemas/PaymentReceipt"
                    }
                  }
                },
                "examples": {
                  "completed": {
                    "value": {
                      "run": {
                        "payload": {
                          "flow": "demo-booking",
                          "status": "completed",
                          "durationMs": 837,
                          "extracted": {
                            "reference": "DB-1FLISL",
                            "guest": "Ada Lovelace <ada@example.com>",
                            "when": "2026-09-01 19:30",
                            "covers": "4",
                            "table": "T07",
                            "errorMessage": null
                          },
                          "screenshots": {
                            "form": "iVBORw0KGgo…",
                            "confirmation": "iVBORw0KGgo…"
                          },
                          "steps": [
                            {
                              "index": 0,
                              "action": "goto",
                              "label": "open the reservation form",
                              "status": "ok",
                              "durationMs": 121
                            }
                          ]
                        },
                        "signature": "…",
                        "algorithm": "HMAC-SHA256"
                      }
                    }
                  },
                  "failed": {
                    "value": {
                      "run": {
                        "payload": {
                          "flow": "demo-booking",
                          "status": "failed",
                          "durationMs": 8351,
                          "error": "step 8 (wait for the confirmation panel) failed: Waiting for selector `#confirmation:not([hidden])` failed",
                          "extracted": {
                            "errorMessage": "No tables available for that name. Please try another time."
                          },
                          "screenshots": {
                            "form": "iVBORw0…",
                            "failure": "iVBORw0…"
                          },
                          "steps": [
                            {
                              "index": 7,
                              "action": "waitFor",
                              "status": "failed",
                              "durationMs": 8003,
                              "error": "Waiting for selector failed"
                            }
                          ]
                        },
                        "signature": "…",
                        "algorithm": "HMAC-SHA256"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Inputs failed validation. FREE — you were not charged.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "INVALID_INPUT",
                  "message": "missing required input \"name\"",
                  "inputs": {},
                  "hint": "Fix the inputs and retry — you have not been charged"
                }
              }
            }
          },
          "402": {
            "$ref": "#/components/responses/PaymentRequired"
          },
          "404": {
            "description": "No such flow. FREE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "FLOW_NOT_FOUND",
                  "flow": "nope",
                  "available": [
                    "demo-booking",
                    "demo-catalog"
                  ]
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "$ref": "#/components/responses/FacilitatorError"
          },
          "503": {
            "description": "No browser could be started. FREE.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "BROWSER_UNAVAILABLE",
                  "message": "No browser could be started, so this flow cannot run. You have not been charged."
                }
              }
            }
          }
        },
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.050000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        }
      }
    },
    "/flows": {
      "get": {
        "tags": [
          "free"
        ],
        "operationId": "listFlows",
        "summary": "Every declared flow with its inputs, extracted fields, price and host allowlist",
        "responses": {
          "200": {
            "description": "The catalogue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "flows": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FlowSummary"
                      }
                    },
                    "loadErrors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "notice": {
                      "type": "string",
                      "description": "Terms-of-service notice about automating third-party sites"
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/verify": {
      "post": {
        "tags": [
          "free"
        ],
        "operationId": "verifyRun",
        "summary": "Verify the HMAC signature on a run record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "payload",
                  "signature"
                ],
                "properties": {
                  "payload": {},
                  "signature": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "valid"
                  ],
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": []
      }
    },
    "/health": {
      "get": {
        "tags": [
          "free"
        ],
        "operationId": "health",
        "summary": "Liveness, payment rails, flow count and browser availability",
        "responses": {
          "200": {
            "description": "Service is up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "service": {
                      "type": "string"
                    },
                    "rails": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Rail"
                      }
                    },
                    "flows": {
                      "type": "integer"
                    },
                    "browserAvailable": {
                      "type": "boolean",
                      "description": "When false, paid runs return a free 503"
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/.well-known/x402": {
      "get": {
        "tags": [
          "free"
        ],
        "operationId": "manifest",
        "summary": "x402 discovery manifest",
        "responses": {
          "200": {
            "description": "Manifest describing paid resources, prices and both payment rails",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "headers": {
      "XPaymentResponse": {
        "description": "Base64 JSON settlement receipt: { success, rail, network, transaction, payer, amount, asset, resource }",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "PaymentRequired": {
        "description": "Payment required. `accepts` lists both rails — pick one, sign it, retry with `X-PAYMENT`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/PaymentRequiredBody"
            },
            "example": {
              "x402Version": 1,
              "error": "X-PAYMENT header is required",
              "accepts": [
                {
                  "scheme": "exact",
                  "network": "base-sepolia",
                  "maxAmountRequired": "50000",
                  "resource": "http://localhost:4034/run/demo-booking",
                  "description": "x402-browser-bridge: POST /run/demo-booking",
                  "mimeType": "application/json",
                  "payTo": "0x40252CFDF8B20Ed757D61ff157719F33Ec332402",
                  "maxTimeoutSeconds": 60,
                  "asset": "0x036CbD53842c5426634e7929541eC2318f3dCF7e",
                  "extra": {
                    "name": "USDC",
                    "version": "2"
                  }
                },
                {
                  "scheme": "exact",
                  "network": "solana",
                  "maxAmountRequired": "50000",
                  "resource": "http://localhost:4034/run/demo-booking",
                  "description": "x402-browser-bridge: POST /run/demo-booking",
                  "mimeType": "application/json",
                  "payTo": "WwwuGbqHrwF5RG89KhUbmRWEvjnRH9k5kVM5p7T3WwW",
                  "maxTimeoutSeconds": 60,
                  "asset": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v",
                  "extra": {
                    "name": "USD Coin",
                    "decimals": 6,
                    "feePayer": "2wKupLR9q6wXYppw8Gr2NvWxKBUqm4PPJKkQfoxHDBg4",
                    "amount": "50000"
                  }
                }
              ]
            }
          }
        }
      },
      "BadRequest": {
        "description": "Malformed request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "Server misconfiguration (e.g. no valid payTo on either rail)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "FacilitatorError": {
        "description": "The payment facilitator could not be reached. Nothing was charged.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "PaymentRequirements": {
        "type": "object",
        "required": [
          "scheme",
          "network",
          "maxAmountRequired",
          "resource",
          "payTo",
          "asset"
        ],
        "properties": {
          "scheme": {
            "type": "string",
            "const": "exact"
          },
          "network": {
            "type": "string",
            "examples": [
              "base-sepolia",
              "base",
              "solana",
              "solana-devnet"
            ]
          },
          "maxAmountRequired": {
            "type": "string",
            "description": "USDC atomic units (6 decimals)"
          },
          "resource": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "payTo": {
            "type": "string"
          },
          "maxTimeoutSeconds": {
            "type": "integer"
          },
          "asset": {
            "type": "string"
          },
          "extra": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "PaymentRequiredBody": {
        "type": "object",
        "required": [
          "x402Version",
          "accepts"
        ],
        "properties": {
          "x402Version": {
            "type": "integer",
            "const": 1
          },
          "error": {
            "type": "string"
          },
          "accepts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentRequirements"
            },
            "minItems": 1
          }
        }
      },
      "PaymentReceipt": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "rail": {
            "type": "string",
            "enum": [
              "evm",
              "solana"
            ]
          },
          "network": {
            "type": "string"
          },
          "transaction": {
            "type": "string"
          },
          "payer": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "asset": {
            "type": "string"
          },
          "resource": {
            "type": "string"
          }
        }
      },
      "Rail": {
        "type": "object",
        "properties": {
          "rail": {
            "type": "string",
            "enum": [
              "evm",
              "solana"
            ]
          },
          "network": {
            "type": "string"
          },
          "payTo": {
            "type": "string"
          },
          "facilitator": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "StepResult": {
        "type": "object",
        "properties": {
          "index": {
            "type": "integer"
          },
          "action": {
            "type": "string",
            "enum": [
              "goto",
              "click",
              "type",
              "select",
              "press",
              "waitFor",
              "waitForText",
              "wait",
              "scroll",
              "screenshot"
            ]
          },
          "label": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "skipped",
              "failed"
            ]
          },
          "durationMs": {
            "type": "number"
          },
          "detail": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "RunResult": {
        "type": "object",
        "properties": {
          "flow": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "failed"
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "durationMs": {
            "type": "number"
          },
          "finalUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "extracted": {
            "type": "object",
            "description": "Values named by the flow's extract block"
          },
          "screenshots": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "description": "base64 PNG, no data-URI prefix"
            }
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StepResult"
            }
          },
          "error": {
            "type": "string",
            "description": "Present when status is failed"
          },
          "inputs": {
            "type": "object"
          }
        }
      },
      "SignedRun": {
        "type": "object",
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/RunResult"
          },
          "signature": {
            "type": "string"
          },
          "algorithm": {
            "type": "string",
            "const": "HMAC-SHA256"
          }
        }
      },
      "FlowInput": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "string",
              "number",
              "boolean"
            ]
          },
          "required": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "default": {},
          "pattern": {
            "type": "string",
            "description": "Anchored regular expression"
          }
        }
      },
      "FlowSummary": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "route": {
            "type": "string"
          },
          "inputs": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/FlowInput"
            }
          },
          "extracts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "screenshots": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "steps": {
            "type": "integer"
          },
          "allowHosts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timeoutMs": {
            "type": "integer"
          },
          "authorized": {
            "type": "boolean",
            "description": "Flow author asserts the target site permits automation. Informational."
          },
          "authorizationNote": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          },
          "flow": {
            "type": "string"
          },
          "available": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "inputs": {
            "type": "object"
          }
        }
      }
    }
  }
}